\name{esotericR-package}
\alias{esotericR-package}
\alias{esotericR}
\docType{package}
\title{
esotericR
}
\description{
Provides code and examples to accompany the esotericR
article series at www.lemnica.com.
}
\details{
\tabular{ll}{
Package: \tab esotericR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-12\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}
\author{
This package is distributed by lemnica, corp. lemnica provides
statistical and technical computing software, training
and support by linking experts with clients.

Maintainer: <jeffrey.ryan@lemnica.com>
}
\references{
\url{http://www.lemnica.com}
%\url{http://www.esotericR.com}
}
\keyword{ package }
